@@ -1,6 +1,6 @@
 module github.com/wwqdrh/tinyagent
 
-go 1.19
+go 1.20
 
 require (
 	github.com/compose-spec/compose-go v1.18.2
