@@ -3,25 +3,20 @@ package srv
 import (
 	"fmt"
 	"strconv"
-	"strings"
 
 	"github.com/wwqdrh/gokit/logger"
-	"github.com/wwqdrh/tinyagent/pkg/swarm"
+	"github.com/wwqdrh/tinyagent/agent/swarm"
 )
 
-type BaseSrvOpt struct {
+type BitnamiMysqlOpt struct {
 	Name     string
 	Image    string
 	Password string // if empty, equals ALLOW_EMPTY_PASSWORD=yes
 	Network  string
 	Ports    map[int]int
-	Envs     []string
-	Args     []string
-	Command  []string
-	Configs  []swarm.ConfigOpt
 }
 
-func (o *BaseSrvOpt) SetOpt(name, value string) {
+func (o *BitnamiMysqlOpt) SetOpt(name, value string) {
 	switch name {
 	case "name":
 		o.Name = value
@@ -32,37 +27,36 @@ func (o *BaseSrvOpt) SetOpt(name, value string) {
 	case "network":
 		o.Network = value
 	case "port":
-		parts := strings.Split(value, ":")
-		if len(parts) != 2 {
-			return
-		}
-		p1, err := strconv.ParseInt(parts[0], 10, 64)
+		v, err := strconv.ParseInt(value, 10, 64)
 		if err != nil {
 			logger.DefaultLogger.Warn(err.Error())
-			return
+		} else {
+			o.Ports[int(v)] = 3306
 		}
-		p2, err := strconv.ParseInt(parts[0], 10, 64)
-		if err != nil {
-			logger.DefaultLogger.Warn(err.Error())
-			return
-		}
-		o.Ports[int(p1)] = int(p2)
-	case "command":
-		o.Command = strings.Split(value, "\n")
 	}
 }
 
-func (o *BaseSrvOpt) Start() error {
+func (o *BitnamiMysqlOpt) envs() (res []string) {
+	if o.Password == "" {
+		res = append(res, "ALLOW_EMPTY_PASSWORD=yes")
+	} else {
+		res = append(res, "MYSQL_ROOT_PASSWORD="+o.Password)
+	}
+	return res
+}
+
+func (o *BitnamiMysqlOpt) Start() error {
+	if o.Image == "" {
+		o.Image = "bitnami/mysql:8.0"
+	}
+
 	res, err := swarm.ServiceCreate(
 		swarm.ServiceOpt{
 			Name:    o.Name,
 			Image:   o.Image,
-			Env:     o.Envs,
+			Env:     o.envs(),
 			Network: o.Network,
 			Ports:   o.Ports,
-			Command: o.Command,
-			Args:    o.Args,
-			Configs: o.Configs,
 		},
 	)
 	if err != nil {
